3799: Extract TransformNodesTest
