+++ common/test/CMakeLists.txt
@@ -112,6 +112,7 @@ set(COMMON_TEST_SOURCE
         "${COMMON_TEST_SOURCE_DIR}/View/SwapNodeContentsTest.cpp"
         "${COMMON_TEST_SOURCE_DIR}/View/TagManagementTest.cpp"
         "${COMMON_TEST_SOURCE_DIR}/View/TextOutputAdapterTest.cpp"
+        "${COMMON_TEST_SOURCE_DIR}/View/TransformNodesTest.cpp"
         "${COMMON_TEST_SOURCE_DIR}/View/UndoTest.cpp"
         "${COMMON_TEST_SOURCE_DIR}/View/UpdateLinkedGroupsHelperTest.cpp"
         "${COMMON_TEST_SOURCE_DIR}/AABBTreeStressTest.cpp"
