+++ common/test/src/View/MapDocumentTest.cpp
@@ -123,236 +123,10 @@ namespace TrenchBroom {
         Quake3MapDocumentTest::Quake3MapDocumentTest() :
         MapDocumentTest{Model::MapFormat::Quake3} {}
 
-        static void checkPlanePointsIntegral(const Model::BrushNode* brushNode) {
-            for (const Model::BrushFace& face : brushNode->brush().faces()) {
-                for (size_t i=0; i<3; i++) {
-                    vm::vec3 point = face.points()[i];
-                    CHECK(pointExactlyIntegral(point));
-                }
-            }
-        }
-
-        static void checkVerticesIntegral(const Model::BrushNode* brushNode) {
-            const Model::Brush& brush = brushNode->brush();
-            for (const Model::BrushVertex* vertex : brush.vertices()) {
-                CHECK(pointExactlyIntegral(vertex->position()));
-            }
-        }
-
-        static void checkBoundsIntegral(const Model::BrushNode* brush) {
-            CHECK(pointExactlyIntegral(brush->logicalBounds().min));
-            CHECK(pointExactlyIntegral(brush->logicalBounds().max));
-        }
-
-        static void checkBrushIntegral(const Model::BrushNode* brush) {
-            checkPlanePointsIntegral(brush);
-            checkVerticesIntegral(brush);
-            checkBoundsIntegral(brush);
-        }
-
         static void setLayerSortIndex(Model::LayerNode& layerNode, int sortIndex) {
             auto layer = layerNode.layer();
             layer.setSortIndex(sortIndex);
             layerNode.setLayer(layer);
-        }
-
-        TEST_CASE_METHOD(MapDocumentTest, "MapDocumentTest.flip") {
-            Model::BrushBuilder builder(document->world()->mapFormat(), document->worldBounds());
-            Model::BrushNode* brushNode1 = new Model::BrushNode(builder.createCuboid(vm::bbox3(vm::vec3(0.0, 0.0, 0.0), vm::vec3(30.0, 31.0, 31.0)), "texture").value());
-            Model::BrushNode* brushNode2 = new Model::BrushNode(builder.createCuboid(vm::bbox3(vm::vec3(30.0, 0.0, 0.0), vm::vec3(31.0, 31.0, 31.0)), "texture").value());
-
-            checkBrushIntegral(brushNode1);
-            checkBrushIntegral(brushNode2);
-
-            addNode(*document, document->parentForNodes(), brushNode1);
-            addNode(*document, document->parentForNodes(), brushNode2);
-
-            std::vector<Model::Node*> brushes;
-            brushes.push_back(brushNode1);
-            brushes.push_back(brushNode2);
-            document->select(brushes);
-
-            const vm::vec3 boundsCenter = document->selectionBounds().center();
-            CHECK(boundsCenter == vm::approx(vm::vec3(15.5, 15.5, 15.5)));
-
-            document->flipObjects(boundsCenter, vm::axis::x);
-
-            checkBrushIntegral(brushNode1);
-            checkBrushIntegral(brushNode2);
-
-            CHECK(brushNode1->logicalBounds() == vm::bbox3(vm::vec3(1.0, 0.0, 0.0), vm::vec3(31.0, 31.0, 31.0)));
-            CHECK(brushNode2->logicalBounds() == vm::bbox3(vm::vec3(0.0, 0.0, 0.0), vm::vec3(1.0, 31.0, 31.0)));
-        }
-
-        TEST_CASE_METHOD(MapDocumentTest, "MapDocumentTest.rotate") {
-            Model::BrushBuilder builder(document->world()->mapFormat(), document->worldBounds());
-            Model::BrushNode* brushNode1 = new Model::BrushNode(builder.createCuboid(vm::bbox3(vm::vec3(0.0, 0.0, 0.0), vm::vec3(30.0, 31.0, 31.0)), "texture").value());
-            Model::BrushNode* brushNode2 = new Model::BrushNode(builder.createCuboid(vm::bbox3(vm::vec3(30.0, 0.0, 0.0), vm::vec3(31.0, 31.0, 31.0)), "texture").value());
-
-            checkBrushIntegral(brushNode1);
-            checkBrushIntegral(brushNode2);
-
-            addNode(*document, document->parentForNodes(), brushNode1);
-            addNode(*document, document->parentForNodes(), brushNode2);
-
-            std::vector<Model::Node*> brushes;
-            brushes.push_back(brushNode1);
-            brushes.push_back(brushNode2);
-            document->select(brushes);
-
-            vm::vec3 boundsCenter = document->selectionBounds().center();
-            CHECK(boundsCenter == vm::vec3(15.5, 15.5, 15.5));
-
-            // 90 degrees CCW about the Z axis through the center of the selection
-            document->rotateObjects(boundsCenter, vm::vec3::pos_z(), vm::to_radians(90.0));
-
-            checkBrushIntegral(brushNode1);
-            checkBrushIntegral(brushNode2);
-
-            const vm::bbox3 brush1ExpectedBounds(vm::vec3(0.0, 0.0, 0.0), vm::vec3(31.0, 30.0, 31.0));
-            const vm::bbox3 brush2ExpectedBounds(vm::vec3(0.0, 30.0, 0.0), vm::vec3(31.0, 31.0, 31.0));
-
-            // these should be exactly integral
-            CHECK(brushNode1->logicalBounds() == brush1ExpectedBounds);
-            CHECK(brushNode2->logicalBounds() == brush2ExpectedBounds);
-        }
-
-        TEST_CASE_METHOD(MapDocumentTest, "MapDocumentTest.shearCube") {
-            const vm::bbox3 initialBBox(vm::vec3(100,100,100), vm::vec3(200,200,200));
-
-            Model::BrushBuilder builder(document->world()->mapFormat(), document->worldBounds());
-            Model::BrushNode* brushNode = new Model::BrushNode(builder.createCuboid(initialBBox, "texture").value());
-
-            addNode(*document, document->parentForNodes(), brushNode);
-            document->select(std::vector<Model::Node*>{brushNode});
-
-            CHECK_THAT(brushNode->brush().vertexPositions(), Catch::UnorderedEquals(std::vector<vm::vec3>{
-                // bottom face
-                {100,100,100},
-                {200,100,100},
-                {200,200,100},
-                {100,200,100},
-                // top face
-                {100,100,200},
-                {200,100,200},
-                {200,200,200},
-                {100,200,200},
-            }));
-
-            // Shear the -Y face by (50, 0, 0). That means the verts with Y=100 will get sheared.
-            CHECK(document->shearObjects(initialBBox, vm::vec3::neg_y(), vm::vec3(50,0,0)));
-
-            CHECK_THAT(brushNode->brush().vertexPositions(), Catch::UnorderedEquals(std::vector<vm::vec3>{
-                // bottom face
-                {150,100,100},
-                {250,100,100},
-                {200,200,100},
-                {100,200,100},
-                // top face
-                {150,100,200},
-                {250,100,200},
-                {200,200,200},
-                {100,200,200},
-            }));
-        }
-
-        TEST_CASE_METHOD(MapDocumentTest, "MapDocumentTest.shearPillar") {
-            const vm::bbox3 initialBBox(vm::vec3(0,0,0), vm::vec3(100,100,400));
-
-            Model::BrushBuilder builder(document->world()->mapFormat(), document->worldBounds());
-            Model::BrushNode* brushNode = new Model::BrushNode(builder.createCuboid(initialBBox, "texture").value());
-
-            addNode(*document, document->parentForNodes(), brushNode);
-            document->select(std::vector<Model::Node*>{brushNode});
-
-            CHECK_THAT(brushNode->brush().vertexPositions(), Catch::UnorderedEquals(std::vector<vm::vec3>{
-                // bottom face
-                {0,  0,  0},
-                {100,0,  0},
-                {100,100,0},
-                {0,  100,0},
-                // top face
-                {0,  0,  400},
-                {100,0,  400},
-                {100,100,400},
-                {0,  100,400},
-            }));
-
-            // Shear the +Z face by (50, 0, 0). That means the verts with Z=400 will get sheared.
-            CHECK(document->shearObjects(initialBBox, vm::vec3::pos_z(), vm::vec3(50,0,0)));
-
-            CHECK_THAT(brushNode->brush().vertexPositions(), Catch::UnorderedEquals(std::vector<vm::vec3>{
-                // bottom face
-                {0,  0,  0},
-                {100,0,  0},
-                {100,100,0},
-                {0,  100,0},
-                // top face
-                {50, 0,  400},
-                {150,0,  400},
-                {150,100,400},
-                {50, 100,400},
-            }));
-        }
-
-        TEST_CASE_METHOD(MapDocumentTest, "MapDocumentTest.scaleObjects") {
-            const vm::bbox3 initialBBox(vm::vec3(-100,-100,-100), vm::vec3(100,100,100));
-            const vm::bbox3 doubleBBox(2.0 * initialBBox.min, 2.0 * initialBBox.max);
-            const vm::bbox3 invalidBBox(vm::vec3(0,-100,-100), vm::vec3(0,100,100));
-
-            Model::BrushBuilder builder(document->world()->mapFormat(), document->worldBounds());
-            Model::BrushNode* brushNode = new Model::BrushNode(builder.createCuboid(initialBBox, "texture").value());
-            const Model::Brush& brush = brushNode->brush();
-
-            addNode(*document, document->parentForNodes(), brushNode);
-            document->select(std::vector<Model::Node*>{brushNode});
-
-            CHECK(brushNode->logicalBounds().size() == vm::vec3(200,200,200));
-            CHECK(brush.face(*brush.findFace(vm::vec3::pos_z())).boundary() == vm::plane3(100.0, vm::vec3::pos_z()));
-
-            // attempting an invalid scale has no effect
-            CHECK_FALSE(document->scaleObjects(initialBBox, invalidBBox));
-            CHECK(brushNode->logicalBounds().size() == vm::vec3(200,200,200));
-            CHECK(brush.face(*brush.findFace(vm::vec3::pos_z())).boundary() == vm::plane3(100.0, vm::vec3::pos_z()));
-
-            CHECK(document->scaleObjects(initialBBox, doubleBBox));
-            CHECK(brushNode->logicalBounds().size() == vm::vec3(400,400,400));
-            CHECK(brush.face(*brush.findFace(vm::vec3::pos_z())).boundary() == vm::plane3(200.0, vm::vec3::pos_z()));
-        }
-
-        TEST_CASE_METHOD(MapDocumentTest, "MapDocumentTest.scaleObjectsInGroup") {
-            const vm::bbox3 initialBBox(vm::vec3(-100, -100, -100), vm::vec3(100, 100, 100));
-            const vm::bbox3 doubleBBox(2.0 * initialBBox.min, 2.0 * initialBBox.max);
-            const vm::bbox3 invalidBBox(vm::vec3(0, -100, -100), vm::vec3(0, 100, 100));
-
-            Model::BrushBuilder builder(document->world()->mapFormat(), document->worldBounds());
-            Model::BrushNode* brushNode = new Model::BrushNode(builder.createCuboid(initialBBox, "texture").value());
-
-            addNode(*document, document->parentForNodes(), brushNode);
-            document->select(std::vector<Model::Node*>{ brushNode });
-            [[maybe_unused]] Model::GroupNode* group = document->groupSelection("my group");
-
-            // attempting an invalid scale has no effect
-            CHECK_FALSE(document->scaleObjects(initialBBox, invalidBBox));
-            CHECK(brushNode->logicalBounds().size() == vm::vec3(200, 200, 200));
-
-            CHECK(document->scaleObjects(initialBBox, doubleBBox));
-            CHECK(brushNode->logicalBounds().size() == vm::vec3(400, 400, 400));
-        }
-
-        TEST_CASE_METHOD(MapDocumentTest, "MapDocumentTest.scaleObjectsWithCenter") {
-            const vm::bbox3 initialBBox(vm::vec3(0,0,0), vm::vec3(100,100,400));
-            const vm::bbox3 expectedBBox(vm::vec3(-50,0,0), vm::vec3(150,100,400));
-
-            Model::BrushBuilder builder(document->world()->mapFormat(), document->worldBounds());
-            Model::BrushNode* brushNode = new Model::BrushNode(builder.createCuboid(initialBBox, "texture").value());
-
-            addNode(*document, document->parentForNodes(), brushNode);
-            document->select(std::vector<Model::Node*>{brushNode});
-
-            const vm::vec3 boundsCenter = initialBBox.center();
-            CHECK(document->scaleObjects(boundsCenter, vm::vec3(2.0, 1.0, 1.0)));
-            CHECK(brushNode->logicalBounds() == expectedBBox);
         }
 
         TEST_CASE_METHOD(MapDocumentTest, "MapDocumentTest.csgConvexMergeBrushes") {
